Add files via upload
It is a simple basic python program that is about a calculator for getting the total price of the addition of different items. In the program, we are taking input from the user and we are adding the price of the items if the "q" button is entered by the user then the total amount of price of the items will be displayed.